More fixes

+++ venice-common/src/main/java/com/linkedin/venice/message/ControlFlagKafkaKey.java
@@ -2,14 +2,14 @@ package com.linkedin.venice.message;
 
 public class ControlFlagKafkaKey extends KafkaKey{
 
-    private int jobId;
+    private long jobId;
 
-    public ControlFlagKafkaKey(OperationType opType, byte[] key, int jobId) {
+    public ControlFlagKafkaKey(OperationType opType, byte[] key, long jobId) {
         super(opType, key);
         this.jobId = jobId;
     }
 
-    public int getJobId(){
+    public long getJobId(){
         return jobId;
     }
 }
